Remove unused method from logger
We don't need to check the validity of the log level, thanks to the
interface design of Logger. Specifically, Logger provides specific
methods for log levels (e.g., info() and debug()), so we remove the risk
and complexity that arise from specification of invalid log levels
(e.g., as an argument to a general method) on the client side.